+++ R/zzz.R
@@ -1,5 +1,6 @@
 #' @importFrom cli cli_alert_warning cli_progress_step cli_progress_done
 #' @importFrom cli col_yellow
+#' @importFrom graphics arrows points text
 #' @importFrom dplyr across all_of
 #' @importFrom dplyr filter join_by left_join mutate rename select
 #' @importFrom stats formula