+++ NEWS.md
@@ -1,3 +1,8 @@
+# classEcoFor (devel)
+
+* Add `triplot_rda()`
+* Use `FLAT_SLOPE` instead of all `SITUATION` values.
+
 # classEcoFor (0.1.0)
 
 * Minimal code to partially reproduce Chalumeau et al.